Add csv to output choices
Allows to print the comparison values in csv format, haven't yet
added functionality to write to a file, but will followup with
another PR for it.. currently this supports print values in a rows.
Will need to make a minor of changes to also support column print.

This also prints metadata towards the end, and fixes the bug with
early exiting when setting output flag